Major revision.  Buy form uses staticfiles, technique-tags, and formsets.
I re-added the static files app, and completely (I think) eliminated the
dependence on technique.mit.edu (including the images).

I added and implemented tags like "gold-price" to enable easy editing
of the patron prices.

Buy page: I implemented the jQuery "Add an older book" and "Remove a book"
buttons to the buy page.  The jQuery copies a hidden, __prefix__ laden example
form and replaces the __prefix__ with the an incrementing number.  Django
handles this as a formset in the purchase.views.buy_form view, although
currently it only prints an acknowledgement.  Next step: ajax up the page so
that changes in the book selections request a "Your Order Is" and "Your Order
Total Is" update.  This puts all the business logic in Django, and none in the
javascript.