Added in basic ajax functionality.
Added dajaxice and dajax libraries, and replicated the previous functionality
of creating the order string and totaling of the old books.  Currently the
patrons and 2012 books are not included in these reports, and the ajax only
runs on a button click.  Next step: add in all the other fields to the form,
and make it only run on change.

Also: everything should be validated through ajax, or else the ajax-added books
vanish.  Either that needs to be fixed or the form should only be able to be
submitted after it is already validated.