Added the current year's book, patrons, and shipping to the ajax.
I hacked the TOTAL_FORMS field to always be 1 or greater, and renamed the
current year's book's fields to fit into the formset.  The ajax pricing logic
now checks if the num_books is greater than 0 as well.

I added an "OrderForm" that tracks the shipping and patron stuff.  I don't know
the best way to get the patron prices... something to look into another time.