Made all field changes trigger validation.
I removed the "Ajaxy Goodness" button and changed the jQuery
selector to update the price and order string on any input
change.  The patron prices are pulled from the Setting table,
which is a little clunky but works well for something that
needs to be accessed in both the template and the validation.

I had to make "0" books an option to let the formset validate
correctly; cludgey, but the "0" books don't show up in the
order string so it's not the end of the world.

Next steps: Validate the purchaser information in the ajax on
a "submit" click.  I'd also like it if the Patron Name and
the shipping address grayed out if no Patron or shipping is
selected (respectively).