Added some more javascript, bundles, messed with the buy page.
Added some javascript validation logic (old books can't have
0 quantity, the newest book isn't in the old books, shipping
address is locked if shipping isn't chosen).  The newest book
thing is bad - it's hard-coded as 2012.  Is there a way to
read that value in to the javascript?

There are freshman and senior bundles on the buy page now,
and the price is stored in the Settings database.  They
run through Ajax too.

Messed with the form styling, probably made it worse.