Corrected the handling of the admin media and templates.
Sometime in the past I completely copied the admin templates
while makine a minor change.  I deleted all the unneccesary
templates and fixed the handling of the admin media.

When I changed the ADMIN_MEDIA prefix, all the little edit
images on the admin site broke; I added those to the static
directory and fixed all the 404 errors that pop up on the
runserver console.