Major overhaul of the orders system.
Books and bundles are now abstracted out - the URL decides which
is rendered on the page.  This vastly simplifies the layout at the
expense of purchaser choice; some version of the old system should
be resurrected as an "Advanced Orders" page.

Currently only 2012 books and the Freshman/Senior bundles work,
but the framework is in place to support any book year.  Patrons
are next in line to be implemented.