Initial Frankenstein grafting-on of localwiki.
There are a lot of extraneous build files and commented-out lines in
settings.py, but both the main site and the wiki function. Both run off of the
django authentication system. I've added a single method decorator to
localwiki/sapling/pages/views.py to restrict viewing any pages to logged-in
users - this should be tested more (it may not cover user pages or detailed
views of pages).