Cleaned up settings.py, and switched to filesystem caching.
The database caching does not play nice with the GIS databases, so I switched
to using the filesystem cache.  The cache folder has to exist so it should
probably be in a /media/cache or just /cache folder within the project.

The wiki maps have Javascript errors and don't work yet.

The wiki detail views ( /_history links and the like) are not password
protected yet.  I'd like a better solution than to edit every view to have the
@login_required method decorator, but I'll go with what works.